+++ src/lib/accessibility/FocusZone/CHANGELOG.md
@@ -5,7 +5,7 @@ This is a list of changes made to this Stardust copy of FocusZone in comparison 
 ### feat(FocusZone): Implement FocusZone into renderComponent [#116](https://github.com/stardust-ui/react/pull/116)
 - Prettier and linting fixes, e.g., removing semicolons, removing underscores from private methods.
 - Moved `IS_FOCUSABLE_ATTRIBUTE` and others to `focusUtilities.ts`.
-- Added prop types.
+- Added prop types, default props, and handled props.
 - Added `preventDefaultWhenHandled` property and method, added to `_onKeyDown`.
 - Renamed boolean callback properties to better reflect their purpose.
 - Renamed `elementType` to `as`.
@@ -13,6 +13,7 @@ This is a list of changes made to this Stardust copy of FocusZone in comparison 
 - Removed `shouldWrapFocus` functionality as it is not necessary for Stardust now and maybe never will.
 - In order to handle custom components in `as`, the `ref` and `_root` were changed to rely on `ReactDOM.findDOMNode` which seems to be necessary because of react-hot-loader. Please see the corresponding [issue #964 at react-hot-loader](https://github.com/gaearon/react-hot-loader/issues/964).
 - Added better typings so that FocusZone passes strict type checks.
+- Fixed `focusLast` mistakes: added it to `FocusZone.types.ts`, fixed a return value, and fixed its comment.
 - Replaced Fabric dependencies accordingly:
     - `BaseComponent` removed.
     - `EventGroup` replaced with vanilla JS approach.